Saving point before refactoring

+++ src/main/java/com/log/gains/day/RequestNewDayDTO.java
@@ -1,4 +1,4 @@
 package com.log.gains.day;
 
-public record RequestNewDayDTO(Double caloriesConsumed, Double weightMeasurement, String date) {
+public record RequestNewDayDTO(Float caloriesConsumed, Float weightMeasurement, String date) {
 }
