+++ src/main/java/com/log/gains/day/RequestedDayUpdateDTO.java
@@ -1,8 +1,3 @@
 package com.log.gains.day;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Component;
-
-import java.time.LocalDate;
-
-public record RequestedDayUpdateDTO(String date, Double weightMeasurement, Double caloriesConsumed){}
+public record RequestedDayUpdateDTO(String date, Float weightMeasurement, Float caloriesConsumed){}
